Introduce a little popover to download the file.
This didn't work, and is more trouble than it's worth. Such a band ida.
And to boot, there's no real way to render the pdf in the first place.
The blob is passed into the DOM and then it makes a url, but there's no
route to serve it from rails.

Stupid idea but committing as "working".

I'm too ill for this.